requirements.txt
pygame
gymnasium
stable-baselines3[extra]
-f https://download.pytorch.org/whl/cpu/torch_stable.html
torch
torchvision
torchaudio
pyyaml
matplotlib
moviepy